requirements.txt
keras==2.9.0
matplotlib==3.5.3
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_learn==1.1.2
scipy==1.9.0
streamlit==1.12.0
tensorflow==2.9.1
